+++ client/packages/ui/src/components/GModal.vue
@@ -1,0 +1,338 @@
+<script setup lang="ts">
+/**
+ * Popup modal component. Offers a main slot, a heading slot, and a footer slot.
+ */
+
+import { FontAwesomeIcon } from "@fortawesome/vue-fontawesome";
+import { watchImmediate } from "@vueuse/core";
+import { faXmark } from "font-awesome-6";
+import { computed, onBeforeUnmount, onMounted, ref } from "vue";
+
+import {
+    type ComponentColor,
+    type ComponentSize,
+    type ComponentSizeClassList,
+    prefix,
+} from "./componentVariants";
+import { useUid } from "../composables/uid";
+
+import GButton from "./GButton.vue";
+import GHeading from "./GHeading.vue";
+
+const HEADING_SIZE_BY_MODAL_SIZE = { small: "sm", medium: "md", large: "lg" } as const;
+
+const props = withDefaults(
+    defineProps<{
+        id?: string;
+        /** Controls if the modal is showing. Syncable */
+        show?: boolean;
+        /** Controls the modals size. If unset, size can be controlled via css `width` and `height` */
+        size?: ComponentSize;
+        /** Makes the modal fill most of the screen with a small margin. If set, ignores other size settings */
+        fullscreen?: boolean;
+        /** Shows confirm an cancel buttons in the footer, and sends out `ok` and `cancel` events */
+        confirm?: boolean;
+        /** Custom text for the Ok confirm button */
+        okText?: string;
+        /** Custom text for the Cancel confirm button */
+        cancelText?: string;
+        /** Renders the footer region, even if confirm is disabled */
+        footer?: boolean;
+        /** Text to display in the title */
+        title?: string;
+        /** Fixes the height of the modal to a pre-set height based on `size` */
+        fixedHeight?: boolean;
+        /** Color of the Ok button */
+        okColor?: ComponentColor;
+        /** Disables the Ok button */
+        okDisabled?: boolean;
+        /** Title to show when the Ok button is disabled */
+        okDisabledTitle?: string;
+        /** When false, keeps the modal open on "ok" */
+        closeOnOk?: boolean;
+        /** Allows content to overflow the modal body (e.g. for dropdowns/selectors inside the modal) */
+        overflowVisible?: boolean;
+    }>(),
+    {
+        id: undefined,
+        show: false,
+        confirm: false,
+        size: undefined,
+        fullscreen: false,
+        okText: undefined,
+        cancelText: undefined,
+        footer: false,
+        title: undefined,
+        fixedHeight: false,
+        okColor: "blue",
+        okDisabled: false,
+        okDisabledTitle: undefined,
+        closeOnOk: true,
+        overflowVisible: false,
+    },
+);
+
+const emit = defineEmits<{
+    (e: "update:show", show: boolean): void;
+    (e: "open"): void;
+    (e: "close"): void;
+    (e: "ok"): void;
+    (e: "cancel"): void;
+}>();
+
+const sizeClass = computed(() => {
+    const classObject: ComponentSizeClassList = {};
+
+    if (props.size) {
+        classObject[prefix(props.size)] = true;
+    }
+
+    return { ...classObject, "g-fixed-height": props.fixedHeight, "g-fullscreen": props.fullscreen };
+});
+
+const dialog = ref<HTMLDialogElement | null>(null);
+
+onMounted(() => {
+    if (dialog.value) {
+        dialog.value.addEventListener("close", onClose);
+    }
+    if (props.show) {
+        showModal();
+    }
+});
+
+onBeforeUnmount(() => {
+    if (dialog.value) {
+        dialog.value.removeEventListener("close", onClose);
+    }
+});
+
+function showModal() {
+    dialog.value?.showModal();
+    onOpen();
+}
+
+let isOk = false;
+
+function hideModal(ok = false) {
+    if (ok && props.closeOnOk === false) {
+        emit("ok");
+    } else {
+        isOk = ok;
+        dialog.value?.close();
+    }
+}
+
+watchImmediate(
+    () => props.show,
+    () => {
+        if (props.show) {
+            showModal();
+        } else {
+            hideModal();
+        }
+    },
+);
+
+function onClickDialog(event: MouseEvent) {
+    if ((event.target as HTMLElement | null)?.tagName === "DIALOG") {
+        const rect = dialog.value?.getBoundingClientRect();
+        const insideDialogX = rect && event.clientX >= rect.left && event.clientX <= rect.right;
+        const insideDialogY = rect && event.clientY >= rect.top && event.clientY <= rect.bottom;
+        const insideDialog = insideDialogX && insideDialogY;
+
+        if (!insideDialog) {
+            hideModal(false);
+        }
+    }
+}
+
+function onOpen() {
+    emit("update:show", true);
+    emit("open");
+
+    isOk = false;
+}
+
+function onClose() {
+    emit("update:show", false);
+    emit("close");
+
+    if (props.confirm && isOk) {
+        emit("ok");
+    } else {
+        emit("cancel");
+    }
+}
+
+const headingSize = computed(() => HEADING_SIZE_BY_MODAL_SIZE[props.size ?? "medium"]);
+
+const uid = useUid("g-modal");
+const currentId = computed(() => props.id ?? uid.value);
+
+defineExpose({ showModal, hideModal });
+</script>
+
+<template>
+    <!-- This is a convenience shortcut for mouse-users to close the dialog, so disabling this warning is fine here -->
+    <!-- eslint-disable-next-line vuejs-accessibility/no-static-element-interactions, vuejs-accessibility/click-events-have-key-events -->
+    <dialog
+        :id="currentId"
+        ref="dialog"
+        class="g-dialog"
+        :class="[sizeClass, { 'g-overflow-visible': props.overflowVisible }]"
+        @click="onClickDialog">
+        <section>
+            <header>
+                <GHeading
+                    v-if="props.title"
+                    h2
+                    :separator="props.size === 'large'"
+                    :size="headingSize"
+                    class="g-modal-title mb-0"
+                    :class="props.size === 'large' ? '' : 'ml-2'">
+                    {{ props.title }}
+                </GHeading>
+
+                <slot name="header"></slot>
+
+                <GButton icon-only class="g-modal-close-button" transparent size="large" @click="hideModal(false)">
+                    <FontAwesomeIcon fixed-width :icon="faXmark" />
+                </GButton>
+            </header>
+
+            <div class="g-modal-content">
+                <slot></slot>
+            </div>
+
+            <footer v-if="props.footer || props.confirm">
+                <div class="g-modal-footer-content">
+                    <slot name="footer"></slot>
+                </div>
+
+                <div v-if="props.confirm" class="g-modal-confirm-buttons">
+                    <GButton @click="hideModal(false)"> {{ props.cancelText ?? "Cancel" }} </GButton>
+                    <GButton
+                        :disabled="okDisabled"
+                        :disabled-title="okDisabledTitle"
+                        :color="props.okColor"
+                        @click="hideModal(true)">
+                        {{ props.okText ?? "Ok" }}
+                    </GButton>
+                </div>
+            </footer>
+        </section>
+    </dialog>
+</template>
+
+<style lang="scss" scoped>
+.g-dialog {
+    background-color: var(--background-color);
+    border-radius: var(--spacing-2);
+    border: none;
+
+    padding: var(--spacing-3);
+
+    max-height: calc(100vh - 4rem);
+
+    section {
+        width: 100%;
+        display: flex;
+        flex-direction: column;
+        gap: var(--spacing-2);
+
+        .g-modal-content {
+            flex-grow: 1;
+            overflow: auto;
+
+            padding: var(--spacing-2);
+            margin: calc(var(--spacing-2) * -1);
+
+            display: flex;
+            flex-direction: column;
+        }
+    }
+
+    &.g-fixed-height {
+        section {
+            height: 100%;
+        }
+    }
+
+    &.g-overflow-visible {
+        overflow: visible;
+
+        section {
+            overflow: visible;
+        }
+
+        .g-modal-content {
+            overflow: visible;
+            max-height: none;
+        }
+    }
+
+    &::backdrop {
+        background-color: var(--color-blue-800);
+        opacity: 0.33;
+    }
+
+    &.g-small {
+        width: 600px;
+
+        &.g-fixed-height {
+            height: 700px;
+        }
+    }
+
+    &.g-medium {
+        width: 900px;
+
+        &.g-fixed-height {
+            height: 750px;
+        }
+    }
+
+    &.g-large {
+        width: 1450px;
+
+        &.g-fixed-height {
+            height: 800px;
+        }
+    }
+
+    &.g-fullscreen {
+        width: calc(100vw - 6rem);
+        height: calc(100vh - 6rem);
+        max-width: calc(100vw - 6rem);
+        max-height: calc(100vh - 6rem);
+
+        section {
+            height: 100%;
+        }
+    }
+
+    header {
+        display: flex;
+        justify-content: flex-end;
+        gap: var(--spacing-1);
+
+        .g-modal-title {
+            flex-grow: 1;
+        }
+    }
+
+    footer {
+        margin: calc(var(--spacing-3) * -1);
+        padding: var(--spacing-3);
+        border-top: 1px solid var(--color-grey-200);
+        display: flex;
+
+        margin-top: var(--spacing-2);
+
+        .g-modal-footer-content {
+            flex-grow: 1;
+        }
+    }
+}
+</style>
