+++ client/packages/ui/src/index.ts
@@ -1,4 +1,42 @@
-// Placeholder barrel for @galaxyproject/galaxy-ui. Components and composables
-// are added here as the extraction from client/src/components/BaseComponents/
-// proceeds.
-export {};
+// @galaxyproject/galaxy-ui public barrel. Internal package, consumed by the
+// main Galaxy client through a Vite source alias (no library build).
+
+export { default as GButton } from "./components/GButton.vue";
+export { default as GButtonGroup } from "./components/GButtonGroup.vue";
+export { default as GCheckbox } from "./components/GCheckbox.vue";
+export { default as GCollapse } from "./components/GCollapse.vue";
+export { default as GHeading } from "./components/GHeading.vue";
+export { default as GLink } from "./components/GLink.vue";
+export { default as GModal } from "./components/GModal.vue";
+export { default as GOverlay } from "./components/GOverlay.vue";
+export { default as GTab } from "./components/GTab.vue";
+export { default as GTabs } from "./components/GTabs.vue";
+export { default as GTip } from "./components/GTip.vue";
+export { default as GTooltip } from "./components/GTooltip.vue";
+
+export { default as GForm } from "./components/Form/GForm.vue";
+export { default as GFormInput } from "./components/Form/GFormInput.vue";
+export { default as GFormLabel } from "./components/Form/GFormLabel.vue";
+
+export {
+    type ColorVariant,
+    type ComponentColor,
+    type ComponentColorClassList,
+    type ComponentSize,
+    type ComponentSizeClassList,
+    type ComponentVariantClassList,
+    prefix,
+} from "./components/componentVariants";
+
+export { useAccessibleHover } from "./composables/accessibleHover";
+export { useClickableElement } from "./composables/clickableElement";
+export { useCurrentTitle } from "./composables/currentTitle";
+export { useMarkdown } from "./composables/markdown";
+export { useResolveElement } from "./composables/resolveElement";
+export { useUid } from "./composables/uid";
+
+export {
+    DEFAULT_TOOLTIP_HOVER_DELAY_MS,
+    INTERACTIVE_POPOVER_CLOSE_DELAY_MS,
+    useDelayedAction,
+} from "./utils/tooltipTiming";
