+++ client/src/components/BaseComponents/Form/GFormInput.vue
@@ -1,71 +1,5 @@
-<script setup lang="ts">
-import { computed, ref } from "vue";
+<script lang="ts">
+import { GFormInput } from "@galaxyproject/galaxy-ui";
 
-const props = defineProps<{
-    value?: string | null;
-}>();
-
-const emit = defineEmits<{
-    (e: "input", value: string | null): void;
-    (e: "keydown", event: KeyboardEvent): void;
-    (e: "blur", event: FocusEvent): void;
-}>();
-
-const inputElement = ref<HTMLInputElement | null>(null);
-
-const inputValue = computed({
-    get() {
-        return props.value;
-    },
-    set(value) {
-        emit("input", value ?? null);
-    },
-});
-
-function focus() {
-    inputElement.value?.focus();
-}
-
-function getInputElement() {
-    return inputElement.value;
-}
-
-defineExpose({
-    focus,
-    getInputElement,
-});
+export default GFormInput;
 </script>
-
-<template>
-    <input
-        ref="inputElement"
-        v-model="inputValue"
-        class="g-form-input"
-        @keydown="(event) => emit('keydown', event)"
-        @blur="(event) => emit('blur', event)" />
-</template>
-
-<style scoped lang="scss">
-.g-form-input {
-    border-radius: var(--spacing-1);
-    border-style: solid;
-    border-width: 1px;
-    border-color: var(--color-grey-400);
-
-    color: var(--color-grey-800);
-
-    padding: var(--spacing-1) var(--spacing-2);
-
-    &:focus {
-        outline: none;
-        box-shadow: 0 0 0 0.2rem rgb(from var(--color-blue-400) r g b / 0.33);
-        z-index: 999;
-    }
-
-    &:focus-visible {
-        outline: none;
-        box-shadow: 0 0 0 0.2rem var(--color-blue-400);
-        z-index: 999;
-    }
-}
-</style>
